requirements.txt
altgraph @ file:///AppleInternal/Library/BuildRoots/bcce998f-ff34-11ef-9d34-f2a857e00a32/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
annotated-types==0.7.0
anyio==4.8.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.2
click==8.1.8
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.18.0
Flask==3.1.0
Flask-SocketIO==5.5.1
fsspec==2025.5.1
future @ file:///AppleInternal/Library/BuildRoots/bcce998f-ff34-11ef-9d34-f2a857e00a32/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
gevent==24.11.1
gevent-websocket==0.10.1
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
llvmlite==0.43.0
macholib @ file:///AppleInternal/Library/BuildRoots/bcce998f-ff34-11ef-9d34-f2a857e00a32/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
MarkupSafe==3.0.2
migrate==0.3.8
more-itertools==10.7.0
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==1.24.4
openai==1.65.2
openai-whisper==20250625
packaging==24.2
pandas==2.2.3
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2025.1
regex==2024.11.6
requests==2.32.4
simple-websocket==1.1.0
six @ file:///AppleInternal/Library/BuildRoots/bcce998f-ff34-11ef-9d34-f2a857e00a32/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
sniffio==1.3.1
sympy==1.14.0
tiktoken==0.9.0
torch==2.2.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.5.0
Werkzeug==3.1.3
wsproto==1.2.0
zipp==3.21.0
zope.event==5.0
zope.interface==7.2
